+++ MCMaps/MCMapsApp.swift
@@ -34,7 +34,8 @@ struct MCMapsApp: App {
     @Environment(\.openURL) private var openURL
 
     @State private var displayCreationWindow = false
-    @State private var proxyMap = MCMapManifest(seed: 0, mcVersion: "1.21", name: "My World", pins: [])
+    @State private var proxyMap = MCMapManifest(
+        name: "My World", worldSettings: MCMapManifestWorldSettings(version: "1.21", seed: 0), pins: [])
 
     init() {
         do {